requirements.txt
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
numpy==1.19.1
pandas==1.1.1
pdfminer.six==20200726
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sortedcontainers==2.2.2
soupsieve==2.0.1
toml==0.10.1
tqdm==4.48.2
Unidecode==1.1.1
urllib3==1.25.10
wrapt==1.12.1
